move push-parameter tests to separate file
less confusing since it makes use of global static initialization
